pyproject.toml
[project]
name = "mingpao-backup"
version = "0.1.0"
description = "Ming Pao Canada HK-GA Wayback Machine archiver"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "internetarchive>=5.7.1",
    "modal>=1.3.0.post1",
    "newspaper4k>=0.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
